+++ src/__tests__/hooks/easy.useSearch.spec.ts
@@ -1,6 +1,6 @@
 import { act, renderHook } from '@testing-library/react';
 
-import { useSearch } from '../../hooks/useSearch.ts';
+import { useSearch } from '../../features/event/model/useSearch.ts';
 import { Event } from '../../types.ts';
 
 const events: Event[] = [
