+++ src/__tests__/hooks/medium.useNotifications.spec.ts
@@ -1,9 +1,9 @@
 import { act, renderHook } from '@testing-library/react';
 
-import { useNotifications } from  '../../hooks/useNotifications.ts';
+import { useNotifications } from '../../features/event/model/useNotifications.ts';
 import { Event } from '../../types.ts';
 
-import { createNotificationMessage } from '../../utils/notificationUtils.ts';
+import { createNotificationMessage } from '@/features/event/lib/notificationUtils.ts';
 import { setupDateMock, cleanupDateMock } from '../dateTimeMockUtils';
