+++ src/__tests__/unit/easy.eventOverlap.spec.ts
@@ -5,7 +5,7 @@ import {
   findOverlappingEvents,
   isOverlapping,
   parseDateTime,
-} from '../../utils/eventOverlap';
+} from '@/features/event/lib/eventOverlap';
 
 describe('parseDateTime', () => {
   it('2024-07-01 14:30을 정확한 Date 객체로 변환한다', () => {
